requirements.txt
pandas==2.2.3
Pillow==11.2.1
plotly==6.1.1
scikit_learn==1.6.1
streamlit==1.45.1
